Added utils module to keep things clean
